Remove last uses of `DreamPath` outside of the compiler
