+++ DMCompiler/Compiler/DM/DMASTHelper.cs
@@ -72,7 +72,7 @@ namespace DMCompiler.Compiler.DM {
             }
 
             public DMASTProcDefinition GetProcByPath(string path) {
-                var h = Hash(new DMASTProcDefinition(Location.Unknown, new OpenDreamShared.Dream.DreamPath(path), new DMASTDefinitionParameter[0], null));
+                var h = Hash(new DMASTProcDefinition(Location.Unknown, new DreamPath(path), new DMASTDefinitionParameter[0], null));
                 return nodes[h][0] as DMASTProcDefinition;
             }
             public void HashDefine(DMASTNode node) {
