+++ DMCompiler/DM/Expressions/Constant.cs
@@ -1,5 +1,4 @@
 using OpenDreamShared.Compiler;
-using OpenDreamShared.Dream;
 using OpenDreamShared.Json;
 using System;
 using System.Collections.Generic;
