+++ DMDisassembler/DMType.cs
@@ -1,27 +1,26 @@
-﻿using OpenDreamShared.Dream;
-using OpenDreamShared.Json;
+﻿using OpenDreamShared.Json;
 using System.Collections.Generic;
 
-namespace DMDisassembler {
-    class DMType {
-        public DreamPath Path;
-        public DreamTypeJson Json;
-        public DMProc InitProc;
-        public Dictionary<string, DMProc> Procs;
+namespace DMDisassembler;
 
-        public DMType(DreamTypeJson json) {
-            Json = json;
-            Path = new DreamPath(Json.Path);
+internal class DMType {
+    public string Path;
+    public DreamTypeJson Json;
+    public DMProc InitProc;
+    public Dictionary<string, DMProc> Procs;
 
-            InitProc = Json.InitProc.HasValue ? Program.Procs[Json.InitProc.Value] : null;
+    public DMType(DreamTypeJson json) {
+        Json = json;
+        Path = Json.Path;
 
-            Procs = new(json.Procs?.Count ?? 0);
-            if (Json.Procs != null) {
-                foreach (List<int> procIds in Json.Procs) {
-                    DMProc proc = Program.Procs[procIds[^1]];
+        InitProc = Json.InitProc.HasValue ? Program.Procs[Json.InitProc.Value] : null;
 
-                    Procs.Add(proc.Name, proc);
-                }
+        Procs = new(json.Procs?.Count ?? 0);
+        if (Json.Procs != null) {
+            foreach (List<int> procIds in Json.Procs) {
+                DMProc proc = Program.Procs[procIds[^1]];
+
+                Procs.Add(proc.Name, proc);
             }
         }
     }
