+++ OpenDreamRuntime/Objects/DreamObject.cs
@@ -425,7 +425,7 @@ namespace OpenDreamRuntime.Objects {
                 return $"{ObjectDefinition.Type}{{name=\"{name}\"}}";
             }
 
-            return ObjectDefinition.Type.ToString();
+            return ObjectDefinition.Type;
         }
     }
 }
