+++ OpenDreamRuntime/Objects/DreamObjectDefinition.cs
@@ -27,7 +27,7 @@ namespace OpenDreamRuntime.Objects {
         public readonly PvsOverrideSystem? PvsOverrideSystem;
 
         public readonly TreeEntry TreeEntry;
-        public DreamPath Type => TreeEntry.Path;
+        public string Type => TreeEntry.Path;
         public DreamObjectDefinition? Parent => TreeEntry.ParentEntry?.ObjectDefinition;
         public int? InitializationProc;
         public bool NoConstructors {
