+++ OpenDreamRuntime/Procs/DMProc.cs
@@ -29,7 +29,7 @@ namespace OpenDreamRuntime.Procs {
 
         private readonly int _maxStackSize;
 
-        public DMProc(int id, DreamPath owningType, ProcDefinitionJson json, string? name, DreamManager dreamManager, AtomManager atomManager, IDreamMapManager dreamMapManager, IDreamDebugManager dreamDebugManager, DreamResourceManager dreamResourceManager, DreamObjectTree objectTree, ProcScheduler procScheduler)
+        public DMProc(int id, TreeEntry owningType, ProcDefinitionJson json, string? name, DreamManager dreamManager, AtomManager atomManager, IDreamMapManager dreamMapManager, IDreamDebugManager dreamDebugManager, DreamResourceManager dreamResourceManager, DreamObjectTree objectTree, ProcScheduler procScheduler)
             : base(id, owningType, name ?? json.Name, null, json.Attributes, GetArgumentNames(json), GetArgumentTypes(json), json.VerbName, json.VerbCategory, json.VerbDesc, json.Invisibility, json.IsVerb) {
             Bytecode = json.Bytecode ?? Array.Empty<byte>();
             LocalNames = json.Locals;
@@ -425,8 +425,8 @@ namespace OpenDreamRuntime.Procs {
         }
 
         public override void AppendStackFrame(StringBuilder builder) {
-            if (Proc.OwningType != DreamPath.Root) {
-                builder.Append(Proc.OwningType.ToString());
+            if (Proc.OwningType != Proc.ObjectTree.Root) {
+                builder.Append(Proc.OwningType);
                 builder.Append('/');
             }
