+++ OpenDreamRuntime/Procs/DebugAdapter/DreamDebugManager.cs
@@ -150,7 +150,7 @@ internal sealed class DreamDebugManager : IDreamDebugManager {
 
         // Check for a function breakpoint
         List<int>? hit = null;
-        if (_possibleFunctionBreakpoints.TryGetValue((state.Proc.OwningType.PathString, state.Proc.Name), out var slot)) {
+        if (_possibleFunctionBreakpoints.TryGetValue((state.Proc.OwningType.Path, state.Proc.Name), out var slot)) {
             foreach (var bp in slot.Breakpoints) {
                 if (TestBreakpoint(bp)) {
                     hit ??= new(1);
@@ -160,7 +160,7 @@ internal sealed class DreamDebugManager : IDreamDebugManager {
         }
 
         if (hit != null) {
-            Output($"Function breakpoint hit at {state.Proc.OwningType.PathString}::{state.Proc.Name}");
+            Output($"Function breakpoint hit at {state.Proc.OwningType.Path}::{state.Proc.Name}");
             Stop(state.Thread, new StoppedEvent {
                 Reason = StoppedEvent.ReasonFunctionBreakpoint,
                 HitBreakpointIds = hit
@@ -422,7 +422,7 @@ internal sealed class DreamDebugManager : IDreamDebugManager {
 
     private IEnumerable<(string Type, string Proc)> IterateProcs() {
         foreach (var proc in _objectTree.Procs) {
-            yield return (proc.OwningType.PathString, proc.Name);
+            yield return (proc.OwningType.Path, proc.Name);
         }
     }
 
@@ -692,7 +692,7 @@ internal sealed class DreamDebugManager : IDreamDebugManager {
     }
 
     private IEnumerable<Variable> ExpandArguments(RequestVariables req, DMProcState dmFrame) {
-        if (dmFrame.Proc.OwningType != OpenDreamShared.Dream.DreamPath.Root) {
+        if (dmFrame.Proc.OwningType != _objectTree.Root) {
             yield return DescribeValue("src", new(dmFrame.Instance));
         }
         yield return DescribeValue("usr", new(dmFrame.Usr));
